Decoder uses modulo arithmetic with dynamic table IDs
The dynamic table is stored in a ring buffer and the absolute IDs
are not based on the total number of inserts.  Instead, these
values are in the range [0, 2 * MaxEntries] and they wrap around.

This means that the decoder can operate forever instead of
erroring out when the number of inserts overflows.