Fix integer state initialization in the decoder
This commit includes a bug fix to the decoder. Previously, the value of `resume` was not reset between state changes and it could lead to incorrect or missing header names.

This commit also includes changes to the the `interop-decode` test program. To increase coverage, there is now a way to limit the number of bytes fed to the decoder at one time. The bug described above was discovered using this rate-limit parameter during testing.

The unit tests now feed the decoder byte by byte.